refactor: replace useHostManagement with useHostState and useContext for improved host management functionality
