+++ src/renderer/src/views/components/AiTab/composables/useContext.ts
@@ -1,14 +1,13 @@
 import { ref, computed, watch, nextTick, onMounted, onUnmounted } from 'vue'
 import type { Ref } from 'vue'
 import debounce from 'lodash/debounce'
-import type { Host, AssetInfo, HostOption, HostItemType } from '../types'
+import type { Host, HostOption, HostItemType } from '../types'
 import { formatHosts, isSwitchAssetType } from '../utils'
 import { isBastionHostType } from '../types'
-import { getBastionHostType } from '../../LeftTab/utils/types'
 import { useSessionState } from './useSessionState'
+import { useHostState } from './useHostState'
 import { focusChatInput } from './useTabManagement'
 import i18n from '@/locales'
-import eventBus from '@/utils/eventBus'
 import { Notice } from '@/views/components/Notice'
 
 /**
@@ -17,77 +16,43 @@ import { Notice } from '@/views/components/Notice'
 const MAX_TARGET_HOSTS = 5
 
 /**
- * Composable for host management
- * Handles host selection, search, pagination loading and other functionalities
+ * Composable for context management (non-singleton)
+ * Handles UI state for host/resource selection, search, and popup positioning
+ * Each component instance gets its own UI state
+ *
+ * Note: This composable will be extended to support other resource types in the future
  */
-export const useHostManagement = () => {
+export const useContext = () => {
   const { t } = i18n.global
 
   const { hosts, chatTypeValue, autoUpdateHost, chatInputValue } = useSessionState()
+  const { getCurentTabAssetInfo } = useHostState()
 
+  // ========== Local UI State (per instance) ==========
   // Popup position in viewport coordinates (used with position: fixed)
   const popupPosition = ref<{ top: number; left: number } | null>(null)
-  // Edit mode popup: keep it hidden until it has a final position.
+  // Edit mode popup: keep it hidden until it has a final position
   const popupReady = ref(false)
   const currentMode = ref<'create' | 'edit'>('create')
 
-  const getCurentTabAssetInfo = async (): Promise<AssetInfo | null> => {
-    const TIMEOUT_MS = 5000
-
-    try {
-      const assetInfo = await new Promise<AssetInfo | null>((resolve, reject) => {
-        const timeout = setTimeout(() => {
-          eventBus.off('assetInfoResult', handleResult)
-          reject(new Error(t('ai.timeoutGettingAssetInfo')))
-        }, TIMEOUT_MS)
-
-        const handleResult = (assetInfo: AssetInfo | null) => {
-          clearTimeout(timeout)
-          eventBus.off('assetInfoResult', handleResult)
-          resolve(assetInfo)
-        }
-        eventBus.on('assetInfoResult', handleResult)
-        eventBus.emit('getActiveTabAssetInfo')
-      })
-
-      if (!assetInfo) {
-        return null
-      }
-
-      // Determine connection type based on assetType
-      // getBastionHostType returns 'jumpserver' | 'qizhi' | null
-      const bastionType = getBastionHostType(assetInfo.assetType)
-      assetInfo.connection = bastionType || 'personal'
-      return assetInfo
-    } catch (error) {
-      console.error('Error getting asset information:', error)
-      return null
-    }
-  }
   const hostSearchInputRef = ref()
   const showHostSelect = ref(false)
-
   const hostOptions = ref<HostOption[]>([])
-
   const hostSearchValue = ref('')
-
   const hostOptionsLoading = ref(false)
   const hostOptionsLimit = 50
-
   const hovered = ref<string | null>(null)
   const keyboardSelectedIndex = ref(-1)
 
   // Track expanded jumpserver nodes (all expanded by default)
   const expandedJumpservers = ref<Set<string>>(new Set())
 
-  // Toggle jumpserver expand/collapse
   const toggleJumpserverExpand = (key: string) => {
     if (expandedJumpservers.value.has(key)) {
       expandedJumpservers.value.delete(key)
     } else {
       expandedJumpservers.value.add(key)
     }
-    // Force reactivity update
     expandedJumpservers.value = new Set(expandedJumpservers.value)
   }
 
@@ -96,7 +61,6 @@ export const useHostManagement = () => {
     const result: HostOption[] = []
 
     for (const item of hostOptions.value) {
-      // Add the item itself
       const isExpanded = expandedJumpservers.value.has(item.key)
       result.push({
         ...item,
@@ -135,7 +99,6 @@ export const useHostManagement = () => {
       return flattenedHostOptions.value
     }
 
-    // When searching, filter items that match
     const result: HostOption[] = []
     for (const item of hostOptions.value) {
       const labelMatches = item.label.toLowerCase().includes(searchTerm)
@@ -148,7 +111,7 @@ export const useHostManagement = () => {
           // Add bastion host node
           result.push({
             ...item,
-            expanded: true // Always expanded when searching
+            expanded: true
           })
           // Add all matching children (or all children if bastion host label matches)
           const childrenToShow = labelMatches ? item.children || [] : matchingChildren
@@ -174,13 +137,9 @@ export const useHostManagement = () => {
     return result
   })
 
-  // Check if host is selected using key for uniqueness (handles duplicate IPs)
   const isHostSelected = (hostOption: HostOption): boolean => {
-    // Use key-based matching to handle duplicate IPs from different sources
     return hosts.value.some((h) => {
-      // Match by uuid for unique identification
       if (h.uuid === hostOption.uuid) return true
-      // Fallback: match by host IP and connection type
       return h.host === hostOption.label && h.connection === hostOption.connect
     })
   }
@@ -214,21 +173,17 @@ export const useHostManagement = () => {
         placement: 'bottomRight'
       })
     } else {
-      // Use key for unique identification when checking existing selection
       const existingIndex = hosts.value.findIndex((h) => h.uuid === item.uuid)
 
       if (existingIndex > -1) {
-        // Remove host if already selected (toggle off)
         hosts.value = hosts.value.filter((_, i) => i !== existingIndex)
       } else {
-        // Check if adding new host would exceed the limit
         let updatedHosts = [...hosts.value]
 
         if (!item.isLocalHost && item.label !== '127.0.0.1') {
           updatedHosts = updatedHosts.filter((h) => h.host !== '127.0.0.1')
         }
 
-        // Validate host count limit before adding
         if (updatedHosts.length >= MAX_TARGET_HOSTS) {
           Notice.open({
             type: 'warning',
@@ -257,6 +212,25 @@ export const useHostManagement = () => {
       hosts.value = hosts.value.filter((_, i) => i !== index)
       autoUpdateHost.value = false
     }
+  }
+
+  const scrollToSelectedItem = () => {
+    nextTick(() => {
+      const selectedItem = document.querySelector('.host-select-item.keyboard-selected') as HTMLElement
+      if (!selectedItem) return
+
+      const hostSelectList = selectedItem.closest('.host-select-list') as HTMLElement
+      if (!hostSelectList) return
+
+      const listRect = hostSelectList.getBoundingClientRect()
+      const itemRect = selectedItem.getBoundingClientRect()
+
+      if (itemRect.top < listRect.top) {
+        hostSelectList.scrollTop -= listRect.top - itemRect.top
+      } else if (itemRect.bottom > listRect.bottom) {
+        hostSelectList.scrollTop += itemRect.bottom - listRect.bottom
+      }
+    })
   }
 
   const handleHostSearchKeyDown = (e: KeyboardEvent, inputValueRef?: Ref<string>) => {
@@ -289,6 +263,7 @@ export const useHostManagement = () => {
           onHostClick(filteredHostOptions.value[keyboardSelectedIndex.value], inputValueRef)
         }
         break
+
       case 'Escape':
         e.preventDefault()
         e.stopPropagation()
@@ -297,31 +272,11 @@ export const useHostManagement = () => {
     }
   }
 
-  const scrollToSelectedItem = () => {
-    nextTick(() => {
-      const selectedItem = document.querySelector('.host-select-item.keyboard-selected') as HTMLElement
-      if (!selectedItem) return
-
-      const hostSelectList = selectedItem.closest('.host-select-list') as HTMLElement
-      if (!hostSelectList) return
-
-      const listRect = hostSelectList.getBoundingClientRect()
-      const itemRect = selectedItem.getBoundingClientRect()
-
-      if (itemRect.top < listRect.top) {
-        hostSelectList.scrollTop -= listRect.top - itemRect.top
-      } else if (itemRect.bottom > listRect.bottom) {
-        hostSelectList.scrollTop += itemRect.bottom - listRect.bottom
-      }
-    })
-  }
-
   const closeHostSelect = () => {
     showHostSelect.value = false
     keyboardSelectedIndex.value = -1
     popupPosition.value = null
     popupReady.value = false
-
     focusChatInput()
   }
 
@@ -339,11 +294,9 @@ export const useHostManagement = () => {
     try {
       const cursorPosition = textarea.selectionStart
 
-      // Create a temporary div to measure cursor position (mirror technique)
       const tempDiv = document.createElement('div')
       const computed = window.getComputedStyle(textarea)
 
-      // Copy all textarea styles for accurate measurement
       const stylesToCopy = [
         'font-family',
         'font-size',
@@ -372,10 +325,9 @@ export const useHostManagement = () => {
       ]
 
       stylesToCopy.forEach((prop) => {
-        tempDiv.style[prop as any] = computed[prop as any]
+        tempDiv.style.setProperty(prop, computed.getPropertyValue(prop))
       })
 
-      // Mirror container setup
       tempDiv.style.position = 'absolute'
       tempDiv.style.visibility = 'hidden'
       tempDiv.style.pointerEvents = 'none'
@@ -386,39 +338,28 @@ export const useHostManagement = () => {
 
       document.body.appendChild(tempDiv)
 
-      // Get text before cursor
       const textBeforeCursor = textarea.value.substring(0, cursorPosition)
       tempDiv.textContent = textBeforeCursor
 
-      // Create a span to mark cursor position
       const span = document.createElement('span')
-      // Use a character or dot to ensure span has height/width
       span.textContent = textarea.value.substring(cursorPosition, cursorPosition + 1) || '.'
       tempDiv.appendChild(span)
 
-      // Get positions relative to the mirror div
       const spanRect = span.getBoundingClientRect()
       const tempDivRect = tempDiv.getBoundingClientRect()
 
-      // Calculate cursor position relative to textarea's content top-left
       const cursorX = spanRect.left - tempDivRect.left
       const cursorY = spanRect.top - tempDivRect.top
 
-      // Calculate line height for vertical offset
       const lineHeightValue = computed.lineHeight
       const lineHeight = lineHeightValue === 'normal' ? parseFloat(computed.fontSize) * 1.2 : parseFloat(lineHeightValue)
 
-      // Clean up mirror element
       document.body.removeChild(tempDiv)
 
-      // Get container and textarea rects to calculate relative positioning
       const textareaRect = textarea.getBoundingClientRect()
-
-      // Caret position in viewport coordinates
       const caretAbsY = textareaRect.top + cursorY - textarea.scrollTop
       const caretAbsX = textareaRect.left + cursorX - textarea.scrollLeft
 
-      // Use chat scroll container as the visible boundary (fallback to window viewport)
       const scrollContainer = textarea.closest('.chat-response-container') as HTMLElement | null
       const scrollRect = scrollContainer?.getBoundingClientRect() ?? {
         top: 0,
@@ -427,25 +368,21 @@ export const useHostManagement = () => {
         bottom: window.innerHeight
       }
 
-      // Popup dimensions (prefer real rendered size; fallback to max constraints)
       const popupEl = document.querySelector('.host-select-popup.is-edit-mode') as HTMLElement | null
       const popupRect = popupEl?.getBoundingClientRect()
       const popupHeight = popupRect?.height ?? 240
       const popupWidth = popupRect?.width ?? 229
       const bufferDistance = 4
 
-      // Below-first decision based on available space inside scrollRect
       const spaceBelow = scrollRect.bottom - (caretAbsY + lineHeight)
       const spaceAbove = caretAbsY - scrollRect.top
 
       const shouldShowBelow = spaceBelow >= popupHeight + bufferDistance
       const shouldShowAbove = !shouldShowBelow && spaceAbove >= popupHeight + bufferDistance
 
-      // Desired popup position in viewport coordinates
       let popupAbsTop = shouldShowAbove ? caretAbsY - popupHeight - bufferDistance : caretAbsY + lineHeight + bufferDistance
       let popupAbsLeft = caretAbsX
 
-      // Clamp into visible scrollRect to reduce clipping
       const minAbsTop = scrollRect.top + 8
       const maxAbsTop = scrollRect.bottom - popupHeight - 8
       popupAbsTop = clamp(popupAbsTop, minAbsTop, maxAbsTop)
@@ -466,7 +403,6 @@ export const useHostManagement = () => {
 
   const calculateCreateModePopupPosition = (triggerEl?: HTMLElement | null) => {
     try {
-      // Find the input container from the trigger element
       let inputContainer: HTMLElement | null = null
       if (triggerEl) {
         inputContainer = triggerEl.closest('.input-send-container') as HTMLElement | null
@@ -481,13 +417,9 @@ export const useHostManagement = () => {
       const marginBottom = 4
       const marginLeft = 8
 
-      // Get popup element to calculate its height for positioning
       const popupEl = document.querySelector('.host-select-popup') as HTMLElement | null
       const popupHeight = popupEl?.getBoundingClientRect().height ?? 240
 
-      // Position popup above the input container
-      // Popup bottom should be at input top - marginBottom
-      // So popup top = input top - popup height - marginBottom
       const top = inputRect.top - popupHeight - marginBottom
       const left = inputRect.left + marginLeft
 
@@ -497,46 +429,6 @@ export const useHostManagement = () => {
       popupPosition.value = null
     }
   }
-
-  const handleInputChange = async (e: Event, mode: 'create' | 'edit' = 'create') => {
-    const value = (e.target as HTMLTextAreaElement).value
-    const textarea = e.target as HTMLTextAreaElement
-
-    currentMode.value = mode
-
-    if (value.endsWith('@')) {
-      if (chatTypeValue.value === 'cmd' || chatTypeValue.value === 'chat') {
-        showHostSelect.value = false
-        popupReady.value = false
-        return
-      }
-
-      showHostSelect.value = true
-      popupReady.value = false
-      hostSearchValue.value = ''
-      await fetchHostOptions('')
-
-      if (mode === 'edit') {
-        calculatePopupPosition(textarea)
-      } else {
-        calculateCreateModePopupPosition(textarea)
-      }
-
-      popupReady.value = true
-      getElement(hostSearchInputRef.value)?.focus?.()
-    } else {
-      showHostSelect.value = false
-      popupPosition.value = null
-    }
-  }
-
-  const debouncedFetchHostOptions = debounce((search: string) => {
-    if (chatTypeValue.value === 'cmd') {
-      fetchHostOptionsForCommandMode(search)
-    } else {
-      fetchHostOptions(search)
-    }
-  }, 300)
 
   const fetchHostOptions = async (search: string) => {
     if (hostOptionsLoading.value) return
@@ -614,6 +506,51 @@ export const useHostManagement = () => {
     }
   }
 
+  const debouncedFetchHostOptions = debounce((search: string) => {
+    if (chatTypeValue.value === 'cmd') {
+      fetchHostOptionsForCommandMode(search)
+    } else {
+      fetchHostOptions(search)
+    }
+  }, 300)
+
+  const getElement = (ref: unknown): HTMLElement | null => {
+    if (!ref) return null
+    return Array.isArray(ref) ? ref[0] || null : (ref as HTMLElement)
+  }
+
+  const handleInputChange = async (e: Event, mode: 'create' | 'edit' = 'create') => {
+    const value = (e.target as HTMLTextAreaElement).value
+    const textarea = e.target as HTMLTextAreaElement
+
+    currentMode.value = mode
+
+    if (value.endsWith('@')) {
+      if (chatTypeValue.value === 'cmd' || chatTypeValue.value === 'chat') {
+        showHostSelect.value = false
+        popupReady.value = false
+        return
+      }
+
+      showHostSelect.value = true
+      popupReady.value = false
+      hostSearchValue.value = ''
+      await fetchHostOptions('')
+
+      if (mode === 'edit') {
+        calculatePopupPosition(textarea)
+      } else {
+        calculateCreateModePopupPosition(textarea)
+      }
+
+      popupReady.value = true
+      getElement(hostSearchInputRef.value)?.focus?.()
+    } else {
+      showHostSelect.value = false
+      popupPosition.value = null
+    }
+  }
+
   const handleAddHostClick = async (triggerEl?: HTMLElement | null) => {
     if (showHostSelect.value) {
       closeHostSelect()
@@ -636,48 +573,6 @@ export const useHostManagement = () => {
       popupReady.value = true
       getElement(hostSearchInputRef.value)?.focus?.()
     })
-  }
-
-  const updateHosts = (hostInfo: { ip: string; uuid: string; connection: string; assetType?: string } | null) => {
-    // Don't update hosts in chat mode
-    if (chatTypeValue.value === 'chat') {
-      hosts.value = []
-      return
-    }
-    if (hostInfo) {
-      if (chatTypeValue.value === 'agent' && isSwitchAssetType(hostInfo.assetType)) {
-        chatTypeValue.value = 'cmd'
-        Notice.open({
-          type: 'info',
-          description: t('ai.switchNotSupportAgent'),
-          placement: 'bottomRight'
-        })
-      }
-      const newHost: Host = {
-        host: hostInfo.ip,
-        uuid: hostInfo.uuid,
-        connection: hostInfo.connection,
-        assetType: hostInfo.assetType
-      }
-      hosts.value = [newHost]
-    } else {
-      hosts.value = []
-    }
-  }
-
-  const updateHostsForCommandMode = async () => {
-    const assetInfo = await getCurentTabAssetInfo()
-
-    if (assetInfo && assetInfo.ip) {
-      updateHosts({
-        ip: assetInfo.ip,
-        uuid: assetInfo.uuid,
-        connection: assetInfo.connection || 'personal',
-        assetType: assetInfo.assetType
-      })
-    } else {
-      updateHosts(null)
-    }
   }
 
   watch(hostSearchValue, (newVal) => {
@@ -703,11 +598,6 @@ export const useHostManagement = () => {
     }
   }
 
-  const getElement = (ref: any): any => {
-    if (!ref) return null
-    return Array.isArray(ref) ? ref[0] || null : ref
-  }
-
   onMounted(() => {
     document.addEventListener('keydown', handleGlobalEscKey)
     document.addEventListener('click', handleGlobalClick)
@@ -719,6 +609,7 @@ export const useHostManagement = () => {
   })
 
   return {
+    // UI state
     showHostSelect,
     hostOptions,
     hostSearchValue,
@@ -727,6 +618,10 @@ export const useHostManagement = () => {
     keyboardSelectedIndex,
     filteredHostOptions,
     hostSearchInputRef,
+    popupPosition,
+    popupReady,
+    currentMode,
+    // UI interaction handlers
     isHostSelected,
     onHostClick,
     removeHost,
@@ -738,12 +633,6 @@ export const useHostManagement = () => {
     fetchHostOptions,
     fetchHostOptionsForCommandMode,
     handleAddHostClick,
-    updateHosts,
-    updateHostsForCommandMode,
-    getCurentTabAssetInfo,
-    toggleJumpserverExpand,
-    popupPosition,
-    popupReady,
-    currentMode
+    toggleJumpserverExpand
   }
 }
